requirements.txt
astroid==2.4.1
autopep8==1.5.4
colorama==0.4.3
console-menu==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.5.2
pyserial==3.4
six==1.14.0
toml==0.10.0
wrapt==1.12.1
